pipeline/requirements.txt
pydub
modelscope
pyannote.audio
pyannote.core
tensorflow
audiosegment
numpy
soundfile
